requirements.txt
django_bootstrap5